Fix fatal errors with query logging.
This code is pretty hacky, but exposes another missing interface.
DebugKit enables query logging for all connections, and this plugin was
blowing up when queries were logged, as it was both missing a logger()
method and the QueryLogger is not a PSR-4 logger.

While this is a bit of a hack it helps unblock other integration work.